+++ lib/src/text_processing/text_preprocessor.dart
@@ -1,0 +1,4 @@
+abstract class TextPreprocessor {
+  String preprocess(String text,
+      {bool caseSensitive = false, List<String>? customStopWords});
+}