+++ lib/src/text_processing/text_processor.dart
@@ -1,0 +1,3 @@
+abstract class TextProcessor {
+  Map<String, int> processText(String text);
+}